+++ fuzz/fuzz_targets/parse_cbor_payload.rs
@@ -1,0 +1,8 @@
+#![no_main]
+
+use libfuzzer_sys::fuzz_target;
+use spex_core::cbor::parse_cbor_payload;
+
+fuzz_target!(|data: &[u8]| {
+    let _ = parse_cbor_payload(data);
+});